Create a FileController Class to deal with file io
Now we need to read the file into memory. I'm going to create my own
FileController that will be responsible
for reading files from the filesystem and saving into memory, and it
will also read files from memory
and save to the file system.

This commit addresses part of issue #9 , but not the whole thing.

Signed-off-by: Isaac A Travers <iat6@humboldt.edu>